Allow redirect_params to contain more (mac source) entries
This fixes a bug, where I could not update/change MAC-addrs
manually via e.g.

 $ sudo ./xdp_prog_user -d veth1 -r veth7 --src-mac ${veth2} --dest-mac ${veth8}
 map dir: /sys/fs/bpf/veth1
 redirect from ifnum=47 to ifnum=54
 WARN: Failed to update bpf map file: err(7):Argument list too long
 can't write iface params

The real bug was that redirect_params is a hash map, with only
a single entry.  It is used in the BPF-code for mapping
MAC src (eth->h_source) to a MAC dest, although current code
doesn't really use result.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>