Fix confusing how _DEBUG is defined in xdp_prog_kern_redirect.c
Debugging bpf_printk will get enabled when defining _DEBUG
even when _DEBUG is defined as "0".

Change define _DEBUG to 1 to make it more clear that
this enables debugging.

Add a comment with #undef _DEBUG as it shows how it should
it is better to change.  Or programmer can uncomment define.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>